Refactor arguments and type tests for auth.* APIs.
